date() typecasting for mktime

+++ src/Resources/contao/classes/ModuleVisitorStatIsotopeProductCounter.php
@@ -47,7 +47,7 @@ class ModuleVisitorStatIsotopeProductCounter extends \BackendModule
         parent::__construct();
 
         $this->today     = date('Y-m-d');
-        $this->yesterday = date('Y-m-d', mktime(0, 0, 0, date("m"), date("d")-1, date("Y")));
+        $this->yesterday = date('Y-m-d', mktime(0, 0, 0, (int) date("m"), (int) date("d")-1, (int) date("Y")));
 
         if (\Database::getInstance()->tableExists('tl_iso_product'))
         {
